+++ src/App.tsx
@@ -10,10 +10,9 @@ function App() {
   return (
     <div className="App">
       <Router>
-          <Route exact path="/" component={TutorialPage}/>
-          <Route exact path="/tt" component={HomePage}/>
-          {/*for link www.x.com/tutorial/1 => id '1' will be passed to TutorialPage component*/}
-          {/*<Route path="/tutorial/:tutorialStep" render={({ match }) => <TutorialPage id={match.params.tutorialStep}/>}/>*/}
+          <Route exact path="/" component={HomePage}/>
+          {/*for url  www.x.com/tutorial/1  => id '1' will be passed to TutorialPage component*/}
+          <Route path="/tutorial/:tutorialStep" render={({ match }) => <TutorialPage id={match.params.tutorialStep}/>}/>
           <Route exact path="/success" component={Completed}/>
       </Router>
     </div>
